# perf: slots on frozen UPat [run_process_replay]